Now comment subscription component is only loaded for logged on users.
