+++ app/javascript/CommentSubscription/index.jsx
@@ -1,1 +1,2 @@
 export * from './CommentSubscription';
+export * from './commentSubscriptionUtilities';
